requirements.txt
# Dependencias del Sistema de Órdenes del Día
# Python 3.8 o superior requerido

# Generación de PDFs
reportlab==4.0.7

# Generación de documentos Word
python-docx==1.1.0

# Lectura de archivos Excel
openpyxl==3.1.5

# Procesamiento de imágenes
Pillow==10.0.0

# Base de datos (incluido en Python estándar)
# sqlite3

# Interfaz gráfica (incluido en Python estándar)
# tkinter